+++ src/ContaoManager/Plugin.php
@@ -8,11 +8,13 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Config\ConfigPluginInterface;
 use ContaoBootstrap\Core\ContaoBootstrapCoreBundle;
 use ContaoBootstrap\Form\ContaoBootstrapFormBundle;
 use Netzmacht\Contao\FormDesigner\NetzmachtContaoFormDesignerBundle;
+use Symfony\Component\Config\Loader\LoaderInterface;
 
-class Plugin implements BundlePluginInterface
+final class Plugin implements BundlePluginInterface, ConfigPluginInterface
 {
     /**
      * {@inheritDoc}
@@ -26,8 +28,16 @@ class Plugin implements BundlePluginInterface
                         ContaoCoreBundle::class,
                         ContaoBootstrapCoreBundle::class,
                         NetzmachtContaoFormDesignerBundle::class,
-                    ]
+                    ],
                 ),
         ];
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig): void
+    {
+        $loader->load(__DIR__ . '/../Resources/config/contao_bootstrap.yaml');
     }
 }