+++ src/FormLayout/DefaultFormLayout.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace ContaoBootstrap\Form\FormLayout;
 
-class DefaultFormLayout extends AbstractBootstrapFormLayout
+final class DefaultFormLayout extends AbstractBootstrapFormLayout
 {
 }